pyproject.toml
[project]
name = "python-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
